+++ tcex/api/tc/v3/tags/mitre_tags.py
@@ -29,13 +29,14 @@ class MitreTag(BaseModel):
 class MitreTags:
     """MitreTags Class"""
 
-    def __init__(self, mitre_tags: dict[str, str]):
+    def __init__(self, mitre_tags: dict[str, str], verbose: bool = False):
         """Initialize instance properties."""
         self._mitre_tags = {id_: MitreTag(id=id_, name=name) for id_, name in mitre_tags.items()}
+        self.verbose = verbose
         self.log = _logger
 
     @cached_property
-    def mitre_tags_name_id(self):
+    def mitre_tags_name_id(self) -> dict[str, MitreTag]:
         """Return a dict of MitreTags keyed by name."""
         mitre_tags = {}
         for tag in self._mitre_tags.values():
@@ -46,29 +47,34 @@ class MitreTags:
 
         return mitre_tags
 
-    def get_by_name(self, name: str) -> str | None:
+    def get_by_name(self, name: str, passthru: bool = True) -> str | None:
         """Return the tag id for the provided name."""
         mitre_tag = self.mitre_tags_name_id.get(name.lower())
         if mitre_tag is None:
-            self.log.warning(f'No Mitre match found for {name}.')
-            return None
+            if self.verbose is True:
+                self.log.warning(f'No Mitre match found for {name}.')
+            return None if passthru is False else name
         return mitre_tag.formatted
 
-    def get_by_id(self, id_: str) -> str:
-        """Return the tag name for the provided id."""
-        mitre_tag = self._mitre_tags.get(id_)
+    def get_by_id(self, id_: str, passthru: bool = True) -> str | None:
+        """Return the tag name for the provided id (e.g., T1000)."""
+        mitre_tag = self._mitre_tags.get(str(id_).upper())
         if mitre_tag is None:
-            self.log.warning(f'No Mitre match found for {id_}, returning id unformatted.')
-            return id_
+            if self.verbose is True:
+                self.log.warning(f'No Mitre match found for {id_}, returning id unformatted.')
+            return None if passthru is False else id_
         return mitre_tag.formatted
 
-    def get_by_id_regex(self, value: str):
+    def get_by_id_regex(self, value: str, passthru: bool = True) -> str | None:
         r"""Get the appropriate MitreTag using the (T\d+(?:\.\d+)?) regex."""
-        matches = re.findall(r'(T\d+(?:\.\d+)?)', value)
+        matches = re.findall(r'([Tt]\d+(?:\.\d+)?)', value)
         if not matches:
-            self.log.warning(f'No Mitre matches found for {value}')
-            return None
+            if self.verbose is True:
+                self.log.warning(f'No Mitre matches found for {value}')
+            return None if passthru is False else value
         if len(matches) > 1:
-            self.log.warning(f'Multiple Mitre matches found for {value}: {matches}')
-            return None
-        return self.get_by_id(matches[0])
+            if self.verbose is True:
+                self.log.warning(f'Multiple Mitre matches found for {value}: {matches}')
+            return None if passthru is False else value
+        mitre_tag = self.get_by_id(matches[0], False)
+        return mitre_tag if mitre_tag is not None else (None if passthru is False else value)